internal/jsonutil: Fix array inside inline fragment edge case.
We might be unmarshaling a single JSON value into multiple GraphQL
fragments or embedded structs. This is represented by the d.vs stacks
of values where to unmarshal.

So, we shouldn't return a fatal error when finding something other than
a slice in one of the places. Instead, skip it. However, if not a single
slice is found in any of the places, then do return an error.

This is similar to how missing struct fields are handled for objects.

Fixes shurcooL/githubql#18.
Updates shurcooL/githubql#15.