Populate partial results on query with some failure.
Sometimes, a part of the GraphQL query can fail, while the rest is
successful. The response will include correct data, and include errors
describing the part of the query that failed. (If the entire query
failed, the response data will simply be null.)

Unmarshal the GraphQL data into supplied structure before returning due
to non-zero GraphQL errors. This way, when there's a partial failure,
it's still possible to access the successful portion of the query.